% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset_calc.R
\name{create_sunset_cal}
\alias{create_sunset_cal}
\title{Creates a sunset calender}
\usage{
create_sunset_cal(date = "2015/01/01", lat = 49.2623, long = -123.2619,
  timezone = "UTC-8", num.days = 365, file = "sunset.csv",
  location = "Wreck Beach, Vancouver, BC V6T")
}
\arguments{
\item{date}{Date at which you want the calendar to start, in yyyy/mm/dd format.}

\item{lat}{Latitude of location (for sunset time calculation)}

\item{long}{Longitude of location (for sunset time calculation, will be negative for continental US)}

\item{timezone}{Timezone of location (for sunset time calculation).}

\item{num.days}{Number of days you want sunset appointments for.}

\item{file}{Filename for outputted .CSV file (to be uploaded to Google Calendar).}

\item{location}{Location of sunset appointment. Will be input into Google Calendar event as the event location.}
}
\value{
.CSV file
}
\description{
Creates a sunset calender
}
\examples{

}

